requirements.txt
-c constraints.txt
coverage
Django==4.0.6
django-db-geventpool
django-environ
djangorestframework
django-filter
gevent
psycopg2-binary
snapshottest==0.5.0
